tools/handbook-ingest/pyproject.toml
[project]
name = "handbook-ingest"
version = "0.1.0"
description = "FAA handbook PDF ingestion pipeline for airboss"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.25,<1.27",
    "beautifulsoup4>=4.12,<5",
    "lxml>=5.3,<6",
    "pillow>=11,<12",
    "click>=8.1,<9",
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
    "ruff==0.6.8",
    "pytest>=8,<9",
]
# Optional: official Anthropic SDK. The pipeline uses raw urllib by default
# so this dependency is opt-in. Install with `pip install -e .[llm]` if you
# want the SDK on PATH; the code falls back to urllib when the SDK is absent.
llm = [
    "anthropic>=0.40,<1",
]

[project.scripts]
handbook-ingest = "ingest.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ingest*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = []
